backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs>=23.0.0
black>=24.0.0
boto3==1.40.41
botocore==1.40.41
cachetools==6.2.0
celery==5.3.4
certifi>=2024.0.0
cffi==2.0.0
charset-normalizer==3.4.3
click>=8.1.0
cryptography==42.0.7
distro==1.9.0
dnspython>=2.6.0,<2.8.0
ecdsa==0.19.1
email-validator==2.3.0
# emergentintegrations - removed (not available on PyPI)
fastapi==0.110.1
fastuuid==0.13.5
filelock==3.19.1
flake8>=7.0.0,<8.0.0
frozenlist==1.8.0
fsspec>=2024.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==1.41.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
jq==1.10.0
jsonschema==4.25.1
jsonschema-specifications>=2023.0.0
litellm==1.77.7
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
multidict==6.6.4
mypy==1.18.2
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
openai==1.99.9
packaging>=24.0
pandas==2.2.2
passlib==1.7.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.4.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle>=2.11.0,<3.0.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyflakes>=3.0.0,<4.0.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio>=0.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
pytz>=2024.0
PyYAML==6.0.3
referencing==0.36.2
regex>=2024.0.0
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.1
redis==5.0.0
rsa==4.9.1
s3transfer==0.14.0
s5cmd==0.2.0
scipy==1.13.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.37.2
stripe==13.0.1
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.1
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata>=2024.0
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.25.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.21.0
zipp==3.23.0
